CreatingTest.js and DeleteTest.js #155
#155 Create folder for creating and deleting
